TX data request for length > 16bits
squash into the previous commit on the memcpy

+++ openair1/PHY/NR_TRANSPORT/nr_dlsch_tools.c
@@ -58,7 +58,7 @@ void nr_fill_dlsch_dl_tti_req(processingData_L1tx_t *msgTx, nfapi_nr_dl_tti_pdsc
   harq->pdu_len = 0;
 }
 
-void nr_fill_dlsch_tx_req(processingData_L1tx_t *msgTx, int idx, uint8_t *pdu, uint16_t pdu_len)
+void nr_fill_dlsch_tx_req(processingData_L1tx_t *msgTx, int idx, uint8_t *pdu, uint32_t pdu_len)
 {
   AssertFatal(pdu != NULL, "pdu is null\n");
